requirements.txt
alembic==1.8.1
attrs==21.4.0
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-Cors==3.0.10
flask-expects-json==1.7.0
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.1
Flask-Migrate==2.6.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.5.1
git-python==1.0.3
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
PyJWT==2.4.0
PyMySQL==1.0.2
pyrsistent==0.18.1
python-environ==0.4.54
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
